Mxe compatability
  * fix warnings.
  * suppress reading and acting upon tx_lock in status
    structure; causes Win-11 segmentation fault.  Most
    likely an indeterminate thread startup conflict.

+++ src/logbook/logsupport.cxx
@@ -595,7 +595,7 @@ std::string adif_record(cQsoRec *rec)
 			snprintf(recfield, sizeof(recfield),
 				"<%s:%lu>",
 				fields[j].name,
-				sFld.length());
+				(unsigned long int)sFld.length());
 			record.append(recfield).append(sFld);
 		}
 	}
